servers/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
perplexityai
